Chagned book types in order_book Updated architecture document

+++ order_book/order_book.v3/order_book.h
@@ -11,8 +11,8 @@
 
 #include <inttypes.h>
 
-const size_t MAX_ORDERS_PER_LEVEL = 10;
-const size_t MAX_ORDER_ID_LEN = 32;
+const uint32_t MAX_ORDERS_PER_LEVEL = 10;
+const uint32_t MAX_ORDER_ID_LEN = 32;
 
 
 struct _order
